Move the row-to-domain conversion and the auction cut into infra
